+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Features
 
-- **packages:** add zhouhaifei-react-demo ([b27155d](https://github.com/air-supply94/zhouhaifei-react-scripts/commit/b27155dd8d4918480a6e2502475c8c9bbfb34cea))
+- **packages:** add zhouhaifei-react-demo ([daaa607](https://github.com/air-supply94/zhouhaifei-react-scripts/commit/daaa607037a07bf9d3c962bffce8d4b5bcae9bf0))
 
 ## [2.14.4](https://github.com/air-supply94/zhouhaifei-react-scripts/compare/v2.14.3...v2.14.4) (2021-08-24)