+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# [0.0.0](https://github.com/air-supply94/zhouhaifei-react-scripts/compare/v2.14.4...v0.0.0) (2021-08-27)
+# [0.0.0](https://github.com/air-supply94/zhouhaifei-react-scripts/compare/v2.15.0...v0.0.0) (2021-10-01)
+
+### Features
+
+- test ([a14b2ef](https://github.com/air-supply94/zhouhaifei-react-scripts/commit/a14b2ef55592605165aa3c019f6a81acfb9c0dd4))
+
+# [2.15.0](https://github.com/air-supply94/zhouhaifei-react-scripts/compare/v2.14.4...v2.15.0) (2021-08-27)
 
 ### Features